We want to use the new peak RAM usage output functionality from develop.

+++ src/openms/include/OpenMS/KERNEL/OnDiscMSExperiment.h
@@ -208,7 +208,7 @@ public:
     */
     OpenMS::Interfaces::SpectrumPtr getSpectrumById(Size id)
     {
-      return indexed_mzml_file_.getSpectrumById(id);
+      return indexed_mzml_file_.getSpectrumById((int)id);
     }
 
     /**
